third_party/blink/web_tests/wpt_internal/shared_storage/run-url-selection-operation-without-add-module.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<script>
'use strict';

promise_test(async t => {
  return promise_rejects_dom(t, "OperationError",
    sharedStorage.runURLSelectionOperation(
      "test-url-selection-operation", ["resources/frame0.html"]));
}, 'runURLSelectionOperation without addModule');

promise_test(async t => {
  try {
      await sharedStorage.runURLSelectionOperation(
          "test-url-selection-operation", ["1", "2", "3", "4", "5", "6", "7", "8", "9"]);
  } catch (e) {
    assert_equals(e.name, 'DataError');
    assert_equals(e.message, 'Length of the \"urls\" parameter is not valid.');
    return;
  }
  assert_unreached("did not reject");
}, 'runURLSelectionOperation with urls array length too big');

promise_test(async t => {
  try {
      await sharedStorage.runURLSelectionOperation(
        "test-url-selection-operation", []);
  } catch (e) {
    assert_equals(e.name, 'DataError');
    assert_equals(e.message, 'Length of the \"urls\" parameter is not valid.');
    return;
  }
  assert_unreached("did not reject");
}, 'runURLSelectionOperation with empty urls array');

</script>
</body>
